Move game history related components to game history feature folder
