+++ frontend/src/features/gameHistory/components/CompletedGameInfo.tsx
