+++ frontend/src/features/gameHistory/components/PageNavigation.tsx
@@ -1,4 +1,4 @@
-import { StateSetterFunction } from "../../../../types/general";
+import { StateSetterFunction } from "../../../types/general";
 import "../../../../styles/features/gameHistory/page-navigation.scss";
 
 type PageNavigationProps = {
